+++ src/gameobject.h
@@ -23,8 +23,8 @@ typedef struct {
 	u8 w;
 	u8 h;
 	BoundBox box;
-	s8 w_offset;
-	s8 h_offset;
+	s16 w_offset;
+	s16 h_offset;
 	u8 anim;
 	u8 health;
 } GameObject;
@@ -32,7 +32,7 @@ typedef struct {
 ////////////////////////////////////////////////////////////////////////////
 // INITIALIZATION
 
-u16 GAMEOBJECT_init(GameObject* const obj, const SpriteDefinition* const sprite, s16 x, s16 y, s8 w_offset, s8 h_offset, u8 pal, u16 ind);
+u16 GAMEOBJECT_init(GameObject* const obj, const SpriteDefinition* const sprite, s16 x, s16 y, s16 w_offset, s16 h_offset, u8 pal, u16 ind);
 
 ////////////////////////////////////////////////////////////////////////////
 // GAME LOOP/LOGIC
